requirements.txt
setuptools

Pillow~=9.3.0
numpy~=1.24.1
pandas~=1.5.2
seaborn~=0.12.2
matplotlib~=3.6.2
ipyplot~=1.1.1
torch~=2.0.0
pyro-ppl